refactor(board): extract CLI business logic into services/boards
Move type-filter matching, payload decoding, and the live items_count
helpers out of src/mondo/cli/board.py into a new src/mondo/services/boards.py,
mirroring the item.py extraction. Moved (renamed _private -> public):
type_matches, decode_json_string_payload, projection_wants_items_count,
fetch_items_count, plus the BoardTypeFilter enum and its server-value map.
Emit/exit-coupled helpers and all @app.command callbacks stay in the CLI.

Renaming _decode_json_string_payload -> decode_json_string_payload made the
output-contract matrix's source annotation stale; update the generator
(scripts/generate_output_contract_matrix.py) and regenerate the two docs.

Behavior-neutral: bodies moved verbatim, call sites rewired, GlobalOpts is a
TYPE_CHECKING-only import. board.py: 1007 -> 929 lines. Full unit suite
(1639) green; ruff/format/mypy clean.
