+++ src/mondo/cli/board.py
@@ -7,7 +7,6 @@ applied client-side after retrieval.
 
 from __future__ import annotations
 
-import json
 import re
 from enum import StrEnum
 from typing import Any
@@ -43,6 +42,13 @@ from mondo.cli._json_flag import parse_json_flag
 from mondo.cli._resolve import resolve_required_id
 from mondo.cli._url import MondayIdParam
 from mondo.cli.context import GlobalOpts
+from mondo.services.boards import (
+    BoardTypeFilter,
+    decode_json_string_payload,
+    fetch_items_count,
+    projection_wants_items_count,
+    type_matches,
+)
 
 app = typer.Typer(
     no_args_is_help=True,
@@ -66,43 +72,6 @@ class BoardState(StrEnum):
 class BoardOrderBy(StrEnum):
     used_at = "used_at"
     created_at = "created_at"
-
-
-class BoardTypeFilter(StrEnum):
-    """`--type` selector on `board list`.
-
-    monday's `boards()` query returns both real boards and workdoc-backing
-    boards (monday models every workdoc as a board with `type=="document"`).
-    The CLI hides docs by default; pass `--type doc` to list only docs, or
-    `--type all` to see everything including non-standard types such as
-    `sub_items_board` and `custom_object`.
-    """
-
-    board = "board"
-    doc = "doc"
-    all = "all"
-
-
-# Mapping from CLI filter → monday's `Board.type` server value.
-_BOARD_TYPE_SERVER_VALUE: dict[BoardTypeFilter, str] = {
-    BoardTypeFilter.board: "board",
-    BoardTypeFilter.doc: "document",
-}
-
-
-def _type_matches(entry: dict[str, Any], type_filter: BoardTypeFilter) -> bool:
-    """Return True when `entry` should pass the `--type` filter.
-
-    Entries cached before schema_version 2 lack `type`; we don't want those
-    to silently disappear under `--type board` (the common default). The
-    schema_version bump forces a one-off refresh so this branch should only
-    matter in the edge case of an in-memory fetch before the cache is warm.
-    Treat missing as `"board"` to keep behavior predictable.
-    """
-    if type_filter is BoardTypeFilter.all:
-        return True
-    observed = entry.get("type") or "board"
-    return observed == _BOARD_TYPE_SERVER_VALUE[type_filter]
 
 
 class BoardAttribute(StrEnum):
@@ -142,16 +111,6 @@ def _resolve_source_workspace(opts: GlobalOpts, board_id: int) -> int | None:
         return int(ws)
     except TypeError, ValueError:
         return None
-
-
-def _decode_json_string_payload(value: Any) -> Any:
-    """Parse monday's legacy stringified-JSON mutation payloads when possible."""
-    if not isinstance(value, str):
-        return value
-    try:
-        return json.loads(value)
-    except json.JSONDecodeError:
-        return value
 
 
 # ----- read commands -----
@@ -367,7 +326,7 @@ def list_cmd(
                         max_items=None,  # client-side filters applied below
                     )
                 )
-                if _type_matches(b, type_filter) and _name_matches(b, needle_lower, pattern)
+                if type_matches(b, type_filter) and _name_matches(b, needle_lower, pattern)
             ]
     except MondoError as e:
         handle_mondo_error_or_exit(e)
@@ -474,7 +433,7 @@ def _list_via_cache(
     if kind is not None:
         entries = [b for b in entries if (b.get("kind") or "") == kind.value]
     if type_filter is not BoardTypeFilter.all:
-        entries = [b for b in entries if _type_matches(b, type_filter)]
+        entries = [b for b in entries if type_matches(b, type_filter)]
     if workspace:
         wanted = {str(w) for w in workspace}
         entries = [b for b in entries if str(b.get("workspace_id") or "") in wanted]
@@ -611,8 +570,8 @@ def get_cmd(
                 )
                 emit_cache_provenance(opts, cached, store=store, explain=explain_cache)
                 board = cached.entries[0] if cached.entries else None
-                if board is not None and _projection_wants_items_count(opts):
-                    items_count = _fetch_items_count(client, board_id)
+                if board is not None and projection_wants_items_count(opts):
+                    items_count = fetch_items_count(client, board_id)
                     board = {**board, "items_count": items_count}
         except NotFoundError:
             board = None
@@ -628,43 +587,6 @@ def get_cmd(
             board = {**board, "url": board_url(get_tenant_slug(client), board_id)}
     board = normalize_board_entry(board)
     opts.emit(board, selected_fields=board_get_fields(with_views=with_views))
-
-
-def _projection_wants_items_count(opts: GlobalOpts) -> bool:
-    """Skip the live `items_count` merge when the caller's projection won't
-    surface it. Default behavior (no `-q`/`--fields`) returns True so the
-    field stays in the emitted payload as it did pre-cache; with an explicit
-    projection we only pay the round-trip if `items_count` actually appears.
-
-    The check is intentionally lenient (substring): a JMESPath expression
-    or comma-separated field list that mentions `items_count` keeps the
-    merge; anything else skips it. False negatives just mean the cached
-    `items_count: null` flows through — accurate per the cache contract.
-    """
-    needle = "items_count"
-    if opts.query and needle in opts.query:
-        return True
-    if opts.fields and needle in opts.fields:
-        return True
-    return opts.query is None and opts.fields is None
-
-
-def _fetch_items_count(client: Any, board_id: int) -> int | None:
-    """One-field live fetch for the volatile items_count field. Used to merge
-    a fresh count onto a `board_details` cache hit so the cache file stays
-    invalidation-free of item writes. Returns None on any unexpected shape."""
-    from mondo.api.queries import BOARD_ITEMS_COUNT
-    from mondo.cli._exec import exec_or_exit
-
-    data = exec_or_exit(client, BOARD_ITEMS_COUNT, {"ids": [board_id]})
-    boards = data.get("boards") or []
-    if not boards:
-        return None
-    raw = boards[0].get("items_count")
-    try:
-        return int(raw) if raw is not None else None
-    except TypeError, ValueError:
-        return None
 
 
 # ----- write commands -----
@@ -753,7 +675,7 @@ def update_cmd(
         {"board": board_id, "attribute": attribute.value, "value": value},
     )
     invalidate_board_caches(opts, board_id)
-    opts.emit(_decode_json_string_payload(data.get("update_board")))
+    opts.emit(decode_json_string_payload(data.get("update_board")))
 
 
 @app.command("set-permission", epilog=epilog_for("board set-permission"))
